[LeeCoder] Modify : 종목 검색 및 구독 해외 데이터로 변경

+++ core/util/src/main/kotlin/com/leecoder/stockchart/util/extension/IntegerExtensions.kt
@@ -2,7 +2,7 @@ package com.leecoder.stockchart.util.extension
 
 import java.text.DecimalFormat
 
-fun Int.toCurrency(): String =
+fun Double.toCurrency(): String =
     try {
         DecimalFormat("#,###").format(this)
     } catch (e: NumberFormatException) {
